Added DOT sample and modified the folder location
