+++ Paragraphs/Ole-image-as-normal-image/.NET/Ole-image-as-normal-image.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30804.86
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Create-Docx-format-Word-document", "Create-Docx-format-Word-document/Create-Docx-format-Word-document.csproj", "{371FD598-B395-4930-AED0-3839092F6FF5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Ole-image-as-normal-image", "Ole-image-as-normal-image/Ole-image-as-normal-image.csproj", "{371FD598-B395-4930-AED0-3839092F6FF5}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
