+++ Word-file-formats/Create-Dot-format-Word-document/.NET/Create-Dot-format-Word-document.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30804.86
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Create-Dot-format-Word-document", "Create-Dot-format-Word-document/Create-Dot-format-Word-document.csproj", "{371FD598-B395-4930-AED0-3839092F6FF5}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{371FD598-B395-4930-AED0-3839092F6FF5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{371FD598-B395-4930-AED0-3839092F6FF5}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{371FD598-B395-4930-AED0-3839092F6FF5}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{371FD598-B395-4930-AED0-3839092F6FF5}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A0DBA025-2B8E-4F35-9527-31803B05C6AC}
+	EndGlobalSection
+EndGlobal
