+++ Word-file-formats/Create-Dot-format-Word-document/.NET/Create-Dot-format-Word-document/Create-Dot-format-Word-document.csproj
@@ -3,7 +3,7 @@
   <PropertyGroup>
     <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
-    <RootNamespace>Create_Docx_format_Word_document</RootNamespace>
+    <RootNamespace>Create_Dot_format_Word_document</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
