--- WordML-conversion/Convert-Word-to-WordML/.NET/Convert-Word-to-WordML.slnx
@@ -1,3 +1,0 @@
-<Solution>
-  <Project Path="Convert-Word-to-WordML/Convert-Word-to-WordML.csproj" />
-</Solution>
