--- WordML-conversion/Convert-WordML-to-Word/.NET/Convert-WordML-to-Word.slnx
@@ -1,3 +1,0 @@
-<Solution>
-  <Project Path="Convert-WordML-to-Word/Convert-WordML-to-Word.csproj" />
-</Solution>
